src/app/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
databricks-sql-connector>=3.6.0
databricks-sdk>=0.38.0
requests>=2.31.0
